kittygram_backend-main/requirements.txt
setuptools>=65.0.0
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
webcolors==1.11.1
Pillow>=10.0.0
gunicorn==21.2.0
dj-database-url==2.2.0
psycopg2-binary==2.9.9
whitenoise==6.7.0
django-cors-headers==4.5.0
ruff>=0.0.200